requirements.txt
Flask
Flask-CORS
requests
psycopg2-binary